+++ presentation/src/commonMain/kotlin/org/moa/moa/di/presentationModule.kt
@@ -1,7 +1,7 @@
 package org.moa.moa.di
 
 import org.koin.dsl.module
-import org.moa.moa.presentation.calendar.CalendarViewModel
+import org.moa.moa.presentation.calendar.calendar.CalendarViewModel
 import org.moa.moa.presentation.calendar.detail.CalendarDetailViewModel
 import org.moa.moa.presentation.record.recorder.RecorderViewModel
 import org.moa.moa.presentation.record.textimage.RecordViewModel
