+++ presentation/src/commonMain/kotlin/org/moa/moa/navigation/calendar/CalendarNavigationGraph.kt
@@ -4,7 +4,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
-import org.moa.moa.presentation.calendar.CalendarScreen
+import org.moa.moa.presentation.calendar.calendar.CalendarScreen
 import org.moa.moa.presentation.calendar.detail.CalendarDetailScreen
 
 fun NavGraphBuilder.calendarNavigationGraph(navController: NavController) {
