+++ presentation/src/commonMain/kotlin/org/moa/moa/presentation/calendar/calendar/component/BottomSheetContentPlaceholder.kt
@@ -1,4 +1,4 @@
-package org.moa.moa.presentation.calendar.component
+package org.moa.moa.presentation.calendar.calendar.component
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
