+++ presentation/src/commonMain/kotlin/org/moa/moa/presentation/calendar/calendar/model/DayInfo.kt
@@ -1,4 +1,4 @@
-package org.moa.moa.presentation.calendar.model
+package org.moa.moa.presentation.calendar.calendar.model
 
 import kotlinx.datetime.LocalDate
 import org.moa.moa.presentation.record.model.Record
